refactor: 사용자 피어리뷰에 대한 클래스 분리
